Worked on the ConfigList Interface and the add/edit/delete config interface
Made the listview functionable and added a fab on the buttom right corner. Also made it so that clicking the listview or the fab will go to the edit and create config interface respectively. In the add/edit/delete interface, added getIntent and getDatafromIntent functions to tell if the user wants to create or edit a game. Added a save and delete icon on the top right corner, and the three inputs for the interface.