Created About Activity Page to have the names of the creators of the application.
All links used is in a listView and clicking on the respective text will send the user to the page where the image, resource is found. Add any resource to the string resource xml. Put Link in string-array "links" and the corresponding name into "link name". Must have the same position/index